improve transparency when things go wrong.  I was using server-side das2 and getting no feedback about what was going on.
svn path=/dasCore/community/autoplot2010/trunk/; revision=5191

+++ dasCore/src/org/das2/DasProperties.java
@@ -266,22 +266,28 @@ public class DasProperties extends Properties {
         try {
             String file= System.getProperty("user.home")+System.getProperty("file.separator")+".das2rc";
             File f= new File(file);
-            
+
             if (f.canRead()) {
                 try {
                     InputStream in= new FileInputStream(f);
                     load(in);
                     in.close();
                 } catch (IOException e) {
+                    e.printStackTrace();
                     org.das2.util.DasExceptionHandler.handle(e);
                 }
             } else {
-                try {
-                    OutputStream out= new FileOutputStream(f);
-                    store(out,"");
-                    out.close();
-                } catch (IOException e) {
-                    org.das2.util.DasExceptionHandler.handle(e);
+                if ( !f.exists() && f.canWrite() ) {
+                    try {
+                        OutputStream out= new FileOutputStream(f);
+                        store(out,"");
+                        out.close();
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                        org.das2.util.DasExceptionHandler.handle(e);
+                    }
+                } else {
+                    System.err.println("Unable to read or write "+file+".  Using defaults.");
                 }
             }
         } catch ( SecurityException ex ) {
